requirements.txt
psycopg2
python-dotenv
bcrypt
